[REL] Migration: stock_picking_ean128_report to 11.0
* [FIX] Update Readme and Manifest
* [FIX] rename xml and .py files according guidelines
* [FIX] identation and some fixes
