+++ stock_picking_ean128_report/__init__.py
@@ -2,4 +2,4 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import wizard
+from . import wizards
