+++ stock_picking_ean128_report/report/stock_picking_print_ean128_report.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo noupdate="1">
+
+    <record id="report_picking_ean128_odt" model="ir.actions.report">
+        <field name="name">Stock Picking EAN128 Report - ODT</field>
+        <field name="model">stock.picking.print_ean128_report</field>
+        <field name="report_name">report_stock_picking_EAN128</field>
+        <field name="report_type">aeroo</field>
+        <field name="in_format">oo-odt</field>
+        <field name="out_format" ref="report_aeroo.report_mimetypes_pdf_odt"/>
+        <field name="report_file">stock_picking_ean128_report/report/stock_picking_EAN128_report.odt</field>
+        <field name="parser_state">default</field>
+        <field name="tml_source">file</field>
+    </record>
+
+    <record id="report_picking_ean128_excel" model="ir.actions.report">
+        <field name="name">Stock Picking EAN128 Report - EXCEL</field>
+        <field name="model">stock.picking.print_ean128_report</field>
+        <field name="report_name">report_stock_picking_EAN128_excel</field>
+        <field name="report_type">aeroo</field>
+        <field name="in_format">oo-ods</field>
+        <field name="out_format" ref="report_aeroo.report_mimetypes_xls_odt"/>
+        <field name="report_file">stock_picking_ean128_report/report/stock_picking_EAN128_report.ods</field>
+        <field name="parser_state">default</field>
+        <field name="tml_source">file</field>
+    </record>
+
+</odoo>
