--- stock_picking_ean128_report/wizard/stock_print_remit_view.xml
@@ -1,119 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-    <data>
-
-
-        <record model="ir.ui.view" id="view_print_ean128_report_detail_form">
-            <field name="name">print_ean128_report_detail.form</field>
-            <field name="model">stock.picking.print_ean128_report_detail</field>
-            <field name="arch" type="xml">
-                <form string="Print Stock Picking Remit">
-                    <group>
-                        <field name="product_id"/>
-                        <field name="product_uom_id"/>
-                        <field name="lot_id"/>
-                        <field name="quantity"/>
-                    </group>
-                </form>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_print_ean128_report_detail_tree">
-            <field name="name">print_ean128_report_detail.tree</field>
-            <field name="model">stock.picking.print_ean128_report_detail</field>
-            <field name="arch" type="xml">
-                <tree string="Print Stock Picking Remit" editable="top" create="0">
-                    <field name="product_id"/>
-                    <field name="product_uom_id"/>
-                    <field name="lot_id"/>
-                    <field name="quantity"/>
-                </tree>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_print_stock_lot_remit_form">
-            <field name="name">print_ean128_lot.form</field>
-            <field name="model">stock.lot.print_ean128_report</field>
-            <field name="arch" type="xml">
-                <form string="Print Stock Lot Remit">
-                    <group>
-                        <field name="product_id"/>
-                        <field name="quantity"/>
-                    </group>
-                    
-                <footer>
-                    <button name="do_print_report" type="object" class="oe_highlight" string="Print"/>
-                    or
-                    <button name="do_print_report_excel" type="object" class="oe_highlight" string="Print List"/>
-                    or
-                    <button string="Cancel" class="oe_link" special="cancel"/>
-                </footer>
-                </form>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_print_stock_picking_remit_form">
-            <field name="name">print_ean128_report.form</field>
-            <field name="model">stock.picking.print_ean128_report</field>
-            <field name="arch" type="xml">
-                <form string="Print Stock Picking Remit">
-                    <group>
-                        <field name="stock_picking_line_ids"/>
-                    </group>
-                    
-                <footer>
-                    <button name="do_print_report" type="object" class="oe_highlight" string="Print"/>
-                    or
-                    <button name="do_print_report_excel" type="object" class="oe_highlight" string="Print List"/>
-                    or
-                    <button string="Cancel" class="oe_link" special="cancel"/>
-                </footer>
-                </form>
-            </field>
-        </record>
-
-        
-
-        <record id="action_stock_print_remit" model="ir.actions.act_window">
-            <field name="name">Print Stock Picking Remit</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">stock.picking.print_ean128_report</field>
-            <field name="src_model">stock.picking</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="view_id" ref="view_print_stock_picking_remit_form"/>
-            <field name="target">new</field>
-        </record>
-
-        <record model="ir.values" id="action_stock_print_remit_values">
-            <field name="model_id" ref="stock.model_stock_picking"/>
-            <field name="name">Print Stock Picking Remit</field>
-            <field name="key2">client_print_multi</field>
-            <field name="value" eval="'ir.actions.act_window,' +str(ref('action_stock_print_remit'))" />
-            <field name="key">action</field>
-            <field name="model">stock.picking</field>
-        </record>
-
-
-        <record id="action_stock_lot_print_remit" model="ir.actions.act_window">
-            <field name="name">Print Stock Lot Remit</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">stock.lot.print_ean128_report</field>
-            <field name="src_model">stock.production.lot</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="view_id" ref="view_print_stock_lot_remit_form"/>
-            <field name="target">new</field>
-        </record>
-
-        <record model="ir.values" id="action_stock_lot_print_remit_values">
-            <field name="model_id" ref="stock.model_stock_picking"/>
-            <field name="name">Print Stock Lot Remit</field>
-            <field name="key2">client_print_multi</field>
-            <field name="value" eval="'ir.actions.act_window,' +str(ref('action_stock_lot_print_remit'))" />
-            <field name="key">action</field>
-            <field name="model">stock.production.lot</field>
-        </record>
-
-    </data>
-</odoo>
