+++ stock_picking_ean128_report/wizards/__init__.py
@@ -2,4 +2,6 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import stock_print_remit
+from . import stock_lot_print_ean128_report
+from . import stock_picking_print_ean128_report
+from . import stock_picking_print_ean128_report_detail
