+++ stock_picking_ean128_report/wizards/stock_lot_print_ean128_report_views.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0"?>
+<odoo>
+
+
+    <record model="ir.ui.view" id="view_print_stock_lot_remit_form">
+        <field name="name">print_ean128_lot.form</field>
+        <field name="model">stock.lot.print_ean128_report</field>
+        <field name="arch" type="xml">
+            <form string="Print Stock Lot Remit">
+                <group>
+                    <field name="product_id"/>
+                    <field name="quantity"/>
+                </group>
+
+                <footer>
+                    <button name="do_print_report" type="object" class="oe_highlight" string="Print"/>
+                or
+                    <button name="do_print_report_excel" type="object" class="oe_highlight" string="Print List"/>
+                or
+                    <button string="Cancel" class="oe_link" special="cancel"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_stock_lot_print_remit" model="ir.actions.act_window">
+        <field name="name">Print Stock Lot Remit</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">stock.lot.print_ean128_report</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form</field>
+        <field name="view_id" ref="view_print_stock_lot_remit_form"/>
+        <field name="target">new</field>
+        <field name="binding_type">report</field>
+        <field name="binding_model_id" ref="stock.model_stock_production_lot"/>
+    </record>
+
+</odoo>
