+++ stock_picking_ean128_report/wizards/stock_picking_print_ean128_report_detail_views.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0"?>
+<odoo>
+    <record model="ir.ui.view" id="view_print_ean128_report_detail_form">
+        <field name="name">print_ean128_report_detail.form</field>
+        <field name="model">stock.picking.print_ean128_report_detail</field>
+        <field name="arch" type="xml">
+            <form string="Print Stock Picking Remit">
+                <group>
+                    <field name="product_id_readonly"/>
+                    <field name="product_uom_id_readonly"/>
+                    <field name="lot_id_readonly"/>
+                    <field name="quantity"/>
+                </group>
+            </form>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_print_ean128_report_detail_tree">
+        <field name="name">print_ean128_report_detail.tree</field>
+        <field name="model">stock.picking.print_ean128_report_detail</field>
+        <field name="arch" type="xml">
+            <tree string="Print Stock Picking Remit" editable="top" create="0">
+                <field name="product_id"/>
+                <field name="product_uom_id"/>
+                <field name="lot_id"/>
+                <field name="quantity"/>
+            </tree>
+        </field>
+    </record>
+</odoo>
