+++ stock_picking_ean128_report/wizards/stock_picking_print_ean128_report_views.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0"?>
+<odoo>
+
+    <record model="ir.ui.view" id="view_print_stock_picking_remit_form">
+        <field name="name">print_ean128_report.form</field>
+        <field name="model">stock.picking.print_ean128_report</field>
+        <field name="arch" type="xml">
+            <form string="Print Stock Picking Remit">
+                <group>
+                    <field name="stock_picking_line_ids"/>
+                </group>
+
+                <footer>
+                    <button name="do_print_report" type="object" class="oe_highlight" string="Print"/>
+                or
+                    <button name="do_print_report_excel" type="object" class="oe_highlight" string="Print List"/>
+                or
+                    <button string="Cancel" class="oe_link" special="cancel"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_stock_print_remit" model="ir.actions.act_window">
+        <field name="name">Print Stock Picking Remit</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">stock.picking.print_ean128_report</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form</field>
+        <field name="view_id" ref="view_print_stock_picking_remit_form"/>
+        <field name="target">new</field>
+        <field name="binding_type">report</field>
+        <field name="binding_model_id" ref="stock.model_stock_picking"/>
+    </record>
+</odoo>
